fix(tracing): resolve set_attribute ambiguity for Clang/libc++
Add explicit static_cast<int64_t> for chrono duration values passed to
span_wrapper::set_attribute() to resolve overload ambiguity between
int64_t and double variants when compiling with Clang and libc++.

The chrono::duration::count() return type (rep) can cause ambiguous
overload resolution with different standard library implementations.

Files fixed:
- message_router.cpp: router.duration_us
- mwl_client.cpp: pacs.duration_ms (4 occurrences)
- hl7_parser.cpp: hl7.parse_time_us

Part of: #144

+++ src/pacs_adapter/mwl_client.cpp
@@ -128,7 +128,7 @@ public:
 
         update_avg_operation_time(elapsed.count());
 
-        span.set_attribute("pacs.duration_ms", elapsed.count());
+        span.set_attribute("pacs.duration_ms", static_cast<int64_t>(elapsed.count()));
         span.set_attribute("pacs.success", true);
 
         return mwl_client::operation_result{
@@ -184,7 +184,7 @@ public:
 
         update_avg_operation_time(elapsed.count());
 
-        span.set_attribute("pacs.duration_ms", elapsed.count());
+        span.set_attribute("pacs.duration_ms", static_cast<int64_t>(elapsed.count()));
         span.set_attribute("pacs.success", true);
 
         return mwl_client::operation_result{
@@ -238,7 +238,7 @@ public:
 
         update_avg_operation_time(elapsed.count());
 
-        span.set_attribute("pacs.duration_ms", elapsed.count());
+        span.set_attribute("pacs.duration_ms", static_cast<int64_t>(elapsed.count()));
         span.set_attribute("pacs.success", true);
 
         return mwl_client::operation_result{
@@ -290,7 +290,7 @@ public:
 
         update_avg_operation_time(elapsed.count());
 
-        span.set_attribute("pacs.duration_ms", elapsed.count());
+        span.set_attribute("pacs.duration_ms", static_cast<int64_t>(elapsed.count()));
         span.set_attribute("pacs.result_count", static_cast<int64_t>(results.size()));
         span.set_attribute("pacs.success", true);
